requirements.txt
asgiref==3.3.1
Django==3.1.3
django-debug-toolbar==3.1.1
flake8==3.8.4
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.4
sqlparse==0.4.1
Pillow=8.0.1
whitenoise=5.2.0
django-crispy-forms=1.10.0


flake8-blind-except
flake8-builtins
flake8-logging-format
flake8-commas
flake8-print
pep8-naming
flake8-cognitive-complexity
flake8-expression-complexity
flake8-functions
flake8-variables-names
flake8-comprehensions
flake8-django
flake8-eradicate
